+++ src/core/client/systems/disable.ts
@@ -3,7 +3,6 @@ import * as native from 'natives';
 import { SHARED_CONFIG } from '../../shared/configurations/shared';
 import { SYSTEM_EVENTS } from '../../shared/enums/System';
 import { Timer } from '../utility/timers';
-import { PushVehicle } from './push';
 import { VehicleController } from './vehicle';
 
 alt.onServer(SYSTEM_EVENTS.TICKS_START, toggleOn);