+++ src/core/client/views/hud.ts
@@ -1,14 +1,16 @@
 import * as alt from 'alt-client';
 import * as native from 'natives';
+import { HUD_COMPONENT } from '../../shared/enums/HudComponents';
 import { SYSTEM_EVENTS } from '../../shared/enums/System';
 import { VEHICLE_STATE } from '../../shared/enums/Vehicle';
+import IClientInteraction from '../../shared/interfaces/IClientInteraction';
+import IHudComponent from '../../shared/interfaces/IHudComponent';
 import { WebViewController } from '../extensions/view2';
 import ViewModel from '../models/ViewModel';
 import { World } from '../systems/world';
 
 const PAGE_NAME = 'Hud';
-const RegisteredComponents: { [key: string]: { callback: Function; msBetweenUpdates: number; lastUpdate?: number } } =
-    {};
+const RegisteredComponents: { [key: string]: IHudComponent } = {};
 
 let interval;
 let isDisabled = false;
@@ -42,35 +44,33 @@ export class HudView implements ViewModel {
         native.displayRadar(false);
     }
 
-    static async setInteractions(_interactions: Array<{ keyPress: string; description: string }>) {
+    static async setInteractions(_interactions: Array<IClientInteraction>) {
         interactions = _interactions;
     }
 
     static async ready() {
-        console.log('HUD Mounted');
-
         if (interval) {
             alt.clearInterval(interval);
         }
 
         // Standard Walking Components
-        HudView.registerComponent('health', HudView.defaultHealthComponent, 50);
-        HudView.registerComponent('armour', HudView.defaultArmourComponent, 50);
-        HudView.registerComponent('cash', HudView.defaultCashComponent, 1000);
-        HudView.registerComponent('bank', HudView.defaultBankComponent, 1000);
-        HudView.registerComponent('time', HudView.defaultTimeComponent, 5000);
-        HudView.registerComponent('water', HudView.defaultWaterComponent, 1000);
-        HudView.registerComponent('food', HudView.defaultFoodComponent, 1000);
-        HudView.registerComponent('interactions', HudView.defaultInteractionsComponent, 500);
+        HudView.registerComponent(HUD_COMPONENT.HEALTH, HudView.defaultHealthComponent, 50);
+        HudView.registerComponent(HUD_COMPONENT.ARMOUR, HudView.defaultArmourComponent, 50);
+        HudView.registerComponent(HUD_COMPONENT.CASH, HudView.defaultCashComponent, 1000);
+        HudView.registerComponent(HUD_COMPONENT.BANK, HudView.defaultBankComponent, 1000);
+        HudView.registerComponent(HUD_COMPONENT.TIME, HudView.defaultTimeComponent, 5000);
+        HudView.registerComponent(HUD_COMPONENT.WATER, HudView.defaultWaterComponent, 1000);
+        HudView.registerComponent(HUD_COMPONENT.FOOD, HudView.defaultFoodComponent, 1000);
+        HudView.registerComponent(HUD_COMPONENT.INTERACTIONS, HudView.defaultInteractionsComponent, 500);
 
         // Vehicle Components
-        HudView.registerComponent('isInVehicle', HudView.defaultIsInVehicleComponent, 1000);
-        HudView.registerComponent('speed', HudView.defaultSpeedComponent, 100);
-        HudView.registerComponent('gear', HudView.defaultGearComponent, 100);
-        HudView.registerComponent('engine', HudView.defaultEngineComponent, 100);
-        HudView.registerComponent('lock', HudView.defaultLockComponent, 100);
-        HudView.registerComponent('metric', HudView.defaultMetricComponent, 2500);
-        HudView.registerComponent('fuel', HudView.defaultFuelComponent, 5000);
+        HudView.registerComponent(HUD_COMPONENT.IS_IN_VEHICLE, HudView.defaultIsInVehicleComponent, 1000);
+        HudView.registerComponent(HUD_COMPONENT.SPEED, HudView.defaultSpeedComponent, 100);
+        HudView.registerComponent(HUD_COMPONENT.GEAR, HudView.defaultGearComponent, 100);
+        HudView.registerComponent(HUD_COMPONENT.ENGINE, HudView.defaultEngineComponent, 100);
+        HudView.registerComponent(HUD_COMPONENT.LOCK, HudView.defaultLockComponent, 100);
+        HudView.registerComponent(HUD_COMPONENT.METRIC, HudView.defaultMetricComponent, 2500);
+        HudView.registerComponent(HUD_COMPONENT.FUEL, HudView.defaultFuelComponent, 5000);
 
         interval = alt.setInterval(HudView.renderComponents, 0);
     }
@@ -92,7 +92,7 @@ export class HudView implements ViewModel {
             }
 
             RegisteredComponents[key].lastUpdate = Date.now() + RegisteredComponents[key].msBetweenUpdates;
-            RegisteredComponents[key].callback();
+            RegisteredComponents[key].callback(key);
         });
     }
 
@@ -107,10 +107,10 @@ export class HudView implements ViewModel {
      * 'HudView.passComponentInfo()'
      * @static
      * @param {number} [msBetweenUpdates=1000]
-     * @param {() => void} callback
+     * @param {(propName: string) => void, msBetweenUpdates: number = 1000} callback
      * @memberof HudView
      */
-    static registerComponent(uid: string, callback: () => void, msBetweenUpdates: number = 1000) {
+    static registerComponent(uid: string, callback: (propName: string) => void, msBetweenUpdates: number = 1000) {
         RegisteredComponents[uid] = {
             callback,
             msBetweenUpdates,
@@ -118,11 +118,28 @@ export class HudView implements ViewModel {
         };
     }
 
+    /**
+     * Remove a component to render on the HUD.
+     * It simply stops it from pushing data the CEF but does not stop it from drawing.
+     * @static
+     * @param {string} uid
+     * @memberof HudView
+     */
+    static unregisterComponent(uid: string) {
+        isDisabled = true;
+
+        if (RegisteredComponents[uid]) {
+            delete RegisteredComponents[uid];
+        }
+
+        isDisabled = false;
+    }
+
     static defaultTimeComponent() {
         HudView.passComponentInfo('time', World.getTimeAsString());
     }
 
-    static defaultFuelComponent() {
+    static defaultFuelComponent(propName: string) {
         if (!alt.Player.local.vehicle) {
             return;
         }
@@ -133,10 +150,10 @@ export class HudView implements ViewModel {
             fuel = alt.Player.local.vehicle.getSyncedMeta(VEHICLE_STATE.FUEL);
         }
 
-        HudView.passComponentInfo('fuel', parseInt(fuel.toFixed(0)));
+        HudView.passComponentInfo(propName, parseInt(fuel.toFixed(0)));
     }
 
-    static defaultSpeedComponent() {
+    static defaultSpeedComponent(propName: string) {
         if (!alt.Player.local.vehicle) {
             return;
         }
@@ -144,75 +161,75 @@ export class HudView implements ViewModel {
         const isMetric = native.getProfileSetting(227);
         const currentSpeed = native.getEntitySpeed(alt.Player.local.vehicle.scriptID);
         const speedCalc = (currentSpeed * (isMetric ? 3.6 : 2.236936)).toFixed(0);
-        HudView.passComponentInfo('speed', parseInt(speedCalc));
+        HudView.passComponentInfo(propName, parseInt(speedCalc));
     }
 
-    static defaultMetricComponent() {
-        HudView.passComponentInfo('isMetric', native.getProfileSetting(227) === 1);
+    static defaultMetricComponent(propName: string) {
+        HudView.passComponentInfo(propName, native.getProfileSetting(227) === 1);
     }
 
-    static defaultIsInVehicleComponent() {
-        HudView.passComponentInfo('isInVehicle', alt.Player.local.vehicle ? true : false);
+    static defaultIsInVehicleComponent(propName: string) {
+        HudView.passComponentInfo(propName, alt.Player.local.vehicle ? true : false);
     }
 
-    static defaultWaterComponent() {
+    static defaultWaterComponent(propName: string) {
         const water = alt.Player.local.meta.water;
-        HudView.passComponentInfo('water', water !== undefined && water !== null ? parseInt(water.toFixed(0)) : 100);
+        HudView.passComponentInfo(propName, water !== undefined && water !== null ? parseInt(water.toFixed(0)) : 100);
     }
 
-    static defaultFoodComponent() {
+    static defaultFoodComponent(propName: string) {
         const food = alt.Player.local.meta.food;
-        HudView.passComponentInfo('food', food !== undefined && food !== null ? parseInt(food.toFixed(0)) : 100);
+        HudView.passComponentInfo(propName, food !== undefined && food !== null ? parseInt(food.toFixed(0)) : 100);
     }
 
-    static defaultCashComponent() {
+    static defaultCashComponent(propName: string) {
         const value = alt.Player.local.meta.cash ? alt.Player.local.meta.cash : 0;
         const fixedValue = parseFloat(value.toFixed(0));
-        HudView.passComponentInfo('cash', fixedValue);
+        HudView.passComponentInfo(propName, fixedValue);
     }
 
-    static defaultBankComponent() {
+    static defaultBankComponent(propName: string) {
         const value = alt.Player.local.meta.bank ? alt.Player.local.meta.bank : 0;
         const fixedValue = parseFloat(value.toFixed(0));
-        HudView.passComponentInfo('bank', fixedValue);
+        HudView.passComponentInfo(propName, fixedValue);
     }
 
-    static defaultGearComponent() {
+    static defaultGearComponent(propName: string) {
         if (!alt.Player.local.vehicle) {
             return;
         }
 
-        HudView.passComponentInfo('gear', alt.Player.local.vehicle.gear);
+        HudView.passComponentInfo(propName, alt.Player.local.vehicle.gear);
     }
 
-    static defaultEngineComponent() {
+    static defaultEngineComponent(propName: string) {
         if (!alt.Player.local.vehicle) {
             return;
         }
 
         const isEngineOn = native.getIsVehicleEngineRunning(alt.Player.local.vehicle.scriptID);
-        HudView.passComponentInfo('engine', isEngineOn);
+        HudView.passComponentInfo(propName, isEngineOn);
     }
 
-    static defaultLockComponent() {
+    static defaultLockComponent(propName: string) {
         if (!alt.Player.local.vehicle) {
             return;
         }
 
         const lockStatus = native.getVehicleDoorLockStatus(alt.Player.local.vehicle.scriptID) === 2;
-        HudView.passComponentInfo('lock', lockStatus);
+        HudView.passComponentInfo(propName, lockStatus);
     }
 
-    static defaultHealthComponent() {
-        HudView.passComponentInfo('health', alt.Player.local.health - 1);
+    static defaultHealthComponent(propName: string) {
+        HudView.passComponentInfo(propName, alt.Player.local.health - 1);
     }
 
-    static defaultArmourComponent() {
-        HudView.passComponentInfo('armour', alt.Player.local.armour);
+    static defaultArmourComponent(propName: string) {
+        HudView.passComponentInfo(propName, alt.Player.local.armour);
     }
 
-    static defaultInteractionsComponent() {
-        HudView.passComponentInfo('interactions', JSON.stringify(interactions), true);
+    static defaultInteractionsComponent(propName) {
+        HudView.passComponentInfo(propName, JSON.stringify(interactions), true);
     }
 }