+++ src/core/server/views/clothing.ts
@@ -20,7 +20,6 @@ import { Vector3 } from '../../shared/interfaces/Vector';
 import { CurrencyTypes } from '../../shared/enums/Currency';
 import { IStreamPolygon } from '../../shared/interfaces/IStreamPolygon';
 import { ServerPolygonController } from '../streamers/polygon';
-import { POLYGON_EVENTS } from '../../shared/enums/PolygonEvents';
 
 // Do not change order
 const icons = [
@@ -106,10 +105,8 @@ class ClothingFunctions {
                     pos: position,
                     vertices: clothingStores[i].vertices,
                     debug: true,
-                    enterEventCall: {
-                        eventName: POLYGON_EVENTS.CLOTHING_SHOP_ENTER,
-                        isServer: true,
-                    },
+                    enterEventCall: ClothingFunctions.enter,
+                    leaveEventCall: ClothingFunctions.leave,
                     maxY: 2.5,
                 };
 
@@ -334,17 +331,21 @@ class ClothingFunctions {
         playerFuncs.emit.sound2D(player, 'item_purchase');
     }
 
-    static enter(player: alt.Player, stream: IStreamPolygon) {
+    static enter<T>(player: T | alt.Player, polygon: IStreamPolygon) {
+        if (!(player instanceof alt.Player)) {
+            return;
+        }
+
         playerFuncs.emit.sound2D(player, 'shop_enter', 0.5);
     }
 
-    static leave(player: alt.Player, stream: IStreamPolygon) {
-        // playerFuncs.emit.message(player, `You have left: ${stream.uid}`);
+    static leave<T>(player: T, polygon: IStreamPolygon) {
+        if (!(player instanceof alt.Player)) {
+            return;
+        }
     }
 }
 
 alt.on(SYSTEM_EVENTS.BOOTUP_ENABLE_ENTRY, ClothingFunctions.init);
 alt.onClient(View_Events_Clothing.Exit, ClothingFunctions.exit);
 alt.onClient(View_Events_Clothing.Purchase, ClothingFunctions.purchase);
-alt.onClient(POLYGON_EVENTS.CLOTHING_SHOP_ENTER, ClothingFunctions.enter);
-alt.onClient(POLYGON_EVENTS.CLOTHING_SHOP_LEAVE, ClothingFunctions.leave);